testapp/requirements.txt
Flask
itsdangerous
Jinja2
MarkupSafe
py-postgresql
uWSGI
